requirements.txt
yfinance
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
statsmodels
cvxpy
fastapi
uvicorn
openai
requests
python-dotenv